Fix progress bar in Home not showing correct progress at first
We can't get progress bar's correct width until its parent view
is laid out, and so the calculation of the progress and applying
that to the width of the foreground was wrong.

This commit changes to set the width constraint on the foreground
using multiplier (current progress / total progress), instead of
calculating the actual width needed to display progress correctly.